+++ packages/core/styles/src/packages/composite/pagination/index.ts
@@ -1,0 +1,5 @@
+import * as styles from "./pagination.style.css";
+
+export const pagination = {
+  classnames: { ...styles },
+};